requirements.txt
# Harness dependencies (host-side scripts: run_episode.py, run_batch.py)
openai>=1.0
httpx>=0.27
pyyaml>=6.0
rich>=13.0
typer>=0.9
